fix toolbarcfg moc include other small changes
git-svn-id: file:///.data/svn_convert/svn@6756 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/branch-QString/sim/cfg.cpp
@@ -834,12 +834,12 @@ EXPORT void saveGeometry(QWidget *w, Geometry geo)
     KWin::WindowInfo info = KWin::windowInfo(w->winId());
     geo[4].asLong() = info.desktop();
     if (info.onAllDesktops())
-        geo[4].asLong() = (unsigned)(-1);
+        geo[4].asLong() = -1;
 #else
     KWin::Info info = KWin::info(w->winId());
     geo[4].asLong() = info.desktop;
     if (info.onAllDesktops)
-        geo[4].asLong() = (unsigned)(-1);
+        geo[4].asLong() = -1;
 #endif
 #endif
 }
